+++ frontend/src/components/prompts/Move.vue
@@ -68,7 +68,7 @@ export default {
       return state.user;
     },
     closeHovers() {
-      return mutations.closeHovers()
+      return mutations.closeHovers();
     },
   },
   methods: {
@@ -91,7 +91,7 @@ export default {
           .then(() => {
             buttons.success("move");
             this.$router.push({ path: this.dest });
-            mutations.setReload(true)
+            mutations.setReload(true);
           })
           .catch((e) => {
             buttons.done("move");
@@ -115,7 +115,7 @@ export default {
             event.preventDefault();
             mutations.closeHovers();
             action(overwrite, rename);
-            mutations.setReload(true)
+            mutations.setReload(true);
           },
         });