+++ frontend/src/store/mutations.js
@@ -1,8 +1,32 @@
 import * as i18n from "@/i18n";
 import { state } from "./state.js";
 import { emitStateChanged } from './eventBus'; // Import the function from eventBus.js
+import { users } from "@/api";
 
 export const mutations = {
+  toggleDarkMode() {
+    mutations.updateUser({ "darkMode": !state.user.darkMode });
+    emitStateChanged();
+  },
+  toggleSidebar() {
+    if (!state.showSidebar && state.user.stickySidebar) {
+      state.user.stickySidebar = false;
+      mutations.updateUser({ "stickySidebar": false }); // turn off sticky when closed
+      return
+    }
+    state.showSidebar = !state.showSidebar;
+    emitStateChanged();
+  },
+  closeSidebar() {
+    if (state.showSidebar) {
+      state.showSidebar = false;
+      emitStateChanged();
+    }
+  },
+  setUpload(value) {
+    state.upload = value;
+    emitStateChanged();
+  },
   setUsage: (value) => {
     state.usage = value;
     emitStateChanged();
@@ -53,12 +77,14 @@ export const mutations = {
     }
 
     let locale = value.locale;
-    if (locale === "") {
-      locale = i18n.detectLocale();
+    if (locale === "") {      
+      value.locale = i18n.detectLocale();
     }
-    i18n.setLocale(locale);
-    i18n.default.locale = locale;
+    let previousUser = state.user
     state.user = value;
+    if (state.user != previousUser && state.user.username != "publicUser") {
+      users.update(state.user);
+    }
     emitStateChanged();
   },
   setJWT: (value) => {
@@ -93,11 +119,15 @@ export const mutations = {
     if (state.user === null) {
       state.user = {};
     }
-    for (let field in value) {
-      if (field === "locale") {
-        i18n.setLocale(value[field]);
-      }
-      state.user[field] = value[field];
+    let previousUser = state.user;
+    state.user = { ...state.user, ...value };
+    if (state.user.locale !== previousUser.locale) {
+      state.user.locale = i18n.detectLocale();
+      i18n.setLocale(state.user.locale);
+      i18n.default.locale = state.user.locale;
+    }
+    if (state.user != previousUser) {
+      users.update(state.user);
     }
     emitStateChanged();
   },
@@ -106,7 +136,6 @@ export const mutations = {
     state.oldReq = state.req;
     state.req = value;
     state.selected = [];
-
     if (!state.req?.items) return;
     state.selected = state.req.items
       .filter((item) => selectedItems.some((rItem) => rItem.url === item.url))