+++ frontend/src/store/state.js
@@ -2,6 +2,7 @@ import { reactive } from 'vue';
 import { detectLocale } from "@/i18n";
 
 export const state = reactive({
+  showSidebar: false,
   usage: {
     used: "0 B",
     total: "0 B",
@@ -9,6 +10,7 @@ export const state = reactive({
   },
   editor: null,
   user: {
+    stickySidebar: false,
     locale: detectLocale(), // Default to the locale from moment
     viewMode: 'mosaic', // Default to mosaic view
     hideDotfiles: false, // Default to false, assuming this is a boolean
@@ -37,18 +39,18 @@ export const state = reactive({
     items: [],
   },
   jwt: "",
-  progress: 0,
   loading: false,
   reload: false,
   selected: [],
   multiple: false,
   upload: {
-    progress: [], // Array of progress values
-    sizes: [],    // Array of sizes
+    uploads: {},
+    queue: [],
+    progress: [],
+    sizes: [],
   },
   prompts: [],
   show: null,
-  showShell: false,
   showConfirm: null,
   route: {},
 });